requirements.txt
# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.0

# Sentiment analysis
textblob>=0.18.0

# Web dashboard
flask>=3.0.0

# Scheduling
apscheduler>=3.10.0

# CLI
click>=8.1.0
rich>=13.0.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
responses>=0.25.0
